renamed all files and made new folders added scrollview to homescreen
