+++ lib/main.dart
@@ -3,12 +3,12 @@ import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import 'package:software_engineering_project_flutter/firebase_options.dart';
 import 'package:firebase_auth/firebase_auth.dart';
-import 'package:software_engineering_project_flutter/pages/home/create_list_screen.dart';
-import 'package:software_engineering_project_flutter/pages/home/create_todo.dart';
-import 'package:software_engineering_project_flutter/pages/home/edit_todo.dart';
-import 'package:software_engineering_project_flutter/pages/home/view_tasks.dart';
+import 'package:software_engineering_project_flutter/pages/home/lists/create_list_screen.dart';
+import 'package:software_engineering_project_flutter/pages/home/tasks/create_task_screen.dart';
+import 'package:software_engineering_project_flutter/pages/home/tasks/edit_task_screen.dart';
+import 'package:software_engineering_project_flutter/pages/home/lists/view_tasks_screen.dart';
 import 'package:software_engineering_project_flutter/wrapper.dart';
-import 'package:software_engineering_project_flutter/pages/home/home.dart';
+import 'package:software_engineering_project_flutter/pages/home/main_screens/home.dart';
 import 'package:software_engineering_project_flutter/pages/authentification/sign_in.dart';
 import 'package:software_engineering_project_flutter/services/authService.dart';
