+++ lib/pages/home/lists/list_of_task_lists_widget.dart
@@ -3,8 +3,8 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
-import 'package:software_engineering_project_flutter/models/taskList.dart';
-import 'package:software_engineering_project_flutter/models/listTileTest.dart';
+import 'package:software_engineering_project_flutter/models/task_list.dart';
+import 'package:software_engineering_project_flutter/models/list_tile.dart';
 import 'package:software_engineering_project_flutter/shared/loading.dart';
 
 class ListOfTaskLists extends StatefulWidget {
