+++ lib/pages/home/tasks/edit_task_screen.dart
@@ -1,34 +1,41 @@
 import 'package:firebase_auth/firebase_auth.dart';
 //import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
+import 'package:software_engineering_project_flutter/models/task.dart';
 import 'package:software_engineering_project_flutter/shared/styles_and_decorations.dart';
 import 'package:intl/intl.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:software_engineering_project_flutter/services/databaseService.dart';
-import 'package:software_engineering_project_flutter/shared/date_time_picker.dart';
+import 'package:software_engineering_project_flutter/shared/date_time_picker_widget.dart';
 import 'package:provider/provider.dart';
 
-class CreateToDo extends StatefulWidget {
-  const CreateToDo({super.key});
+class EditTodo extends StatefulWidget {
+  
+  final Task task; 
+  const EditTodo({required this.task, super.key});
 
   @override
-  State<CreateToDo> createState() => _CreateToDoState();
+  State<EditTodo> createState() => _EditTodoState();
 }
 
-class _CreateToDoState extends State<CreateToDo> {
+class _EditTodoState extends State<EditTodo> {
+
+  late Task task = widget.task;
   final _formKey = GlobalKey<FormState>();
 
   //Felder eines ToDos
-  String title = '';
-  String note = '';
-  String list = 'default';
-  String priority = 'no priority';
-  DateTime dateAndTime = DateTime.fromMillisecondsSinceEpoch(0); //soll nicht null sein
+  late String title = task.description;
+  late String note = task.note;
+  late String list = 'default';
+  late String priority = task.priority;
+  late DateTime maturityDate = task.maturityDate; //soll nicht null sein
 
   //Listen für die Dropdowns
   List<String> categories = ['Arbeit', 'Schule', 'Haushalt'];
   List<String> priorities = ['Hoch', 'Mittel', 'Niedrig'];
   List<DocumentReference> lists = [];
+
+
 
   @override
   Widget build(BuildContext context) {
@@ -40,7 +47,7 @@ class _CreateToDoState extends State<CreateToDo> {
       appBar: AppBar(
         title: Center(
             child: Text(
-          'To-Do erstellen',
+          'To-Do bearbeiten',
           style: standardAppBarTextDecoration,
         )),
         backgroundColor: const Color.fromRGBO(101, 167, 101, 1),
@@ -75,7 +82,7 @@ class _CreateToDoState extends State<CreateToDo> {
                             children: <Widget>[
                               const SizedBox(
                                 child: Icon(
-                                  Icons.add,
+                                  Icons.circle_outlined,
                                   color: Colors.white,
                                   size: 40.0,
                                 ),
@@ -90,7 +97,7 @@ class _CreateToDoState extends State<CreateToDo> {
                                   //Bezeichnung eingeben
                                   child: TextFormField(
                                     style: const TextStyle(color: Colors.white),
-                                    initialValue: "",
+                                    initialValue: title,
                                     validator: (value) {
                                       if (value!.isEmpty) {
                                         return 'Bitte eine Bezeichnung eingeben';
@@ -178,6 +185,7 @@ class _CreateToDoState extends State<CreateToDo> {
                                 elevation: 8,
                                 shadowColor: const Color(0xFF212121),
                                 child: DropdownButtonFormField<String>(
+                                  value: task.priority == 'no priority' ? null : task.priority,
                                   decoration: textInputDecoration.copyWith(
                                       hintText: 'Priorität'),
                                   dropdownColor:
@@ -232,10 +240,10 @@ class _CreateToDoState extends State<CreateToDo> {
                                             await showDateTimePicker(
                                                 context: context) ?? DateTime.fromMicrosecondsSinceEpoch(0);
                                         setState(() {
-                                          dateAndTime = pickedDate;
+                                          maturityDate = pickedDate;
                                         });
                                       },
-                                      child: dateAndTime == DateTime.fromMillisecondsSinceEpoch(0)
+                                      child: maturityDate == DateTime.fromMillisecondsSinceEpoch(0)
                                           ? const Align(
                                               alignment: Alignment.centerLeft,
                                               child: Text(
@@ -247,7 +255,7 @@ class _CreateToDoState extends State<CreateToDo> {
                                               ),
                                             )
                                           : Text(
-                                              '${DateFormat('dd.MM.yyyy').format(dateAndTime)} ${dateAndTime!.hour.toString().padLeft(2, '0')}:${dateAndTime.minute.toString().padLeft(2, '0')}'),
+                                              '${DateFormat('dd.MM.yyyy').format(maturityDate)} ${maturityDate.hour.toString().padLeft(2, '0')}:${maturityDate.minute.toString().padLeft(2, '0')}'),
                                     )),
                               ]),
                           const SizedBox(
@@ -270,7 +278,7 @@ class _CreateToDoState extends State<CreateToDo> {
                                   validator: (value) => value!.length > 300
                                       ? 'Notiz darf nicht länger als 300 Zeichen sein'
                                       : null,
-                                  initialValue: "",
+                                  initialValue: note,
                                   decoration: textInputDecoration.copyWith(
                                       hintText: 'Notiz'),
                                   onChanged: (value) => setState(() {
@@ -305,11 +313,10 @@ class _CreateToDoState extends State<CreateToDo> {
                                 //Erstellen Button
                                 child: ElevatedButton(
                                   style: buttonStyleDecorationcolorchange,
-                                  child: const Text('Erstellen'),
+                                  child: const Text('Speichern'),
                                   onPressed: () {
                                     if (_formKey.currentState!.validate()) {
-                                      _database.addTask(title, note,
-                                          dateAndTime, false, priority, lists, false);
+                                      _database.editTask(title, note, maturityDate, false, priority, null, false, task.taskReference);
                                       Navigator.pop(context);
                                     }
                                   },
