+++ lib/wrapper.dart
@@ -1,7 +1,7 @@
 import 'package:provider/provider.dart';
 import 'package:software_engineering_project_flutter/pages/authentification/authenticate.dart';
 import 'package:software_engineering_project_flutter/pages/authentification/sign_in.dart';
-import 'package:software_engineering_project_flutter/pages/home/home.dart';
+import 'package:software_engineering_project_flutter/pages/home/main_screens/home.dart';
 import 'package:flutter/material.dart';
 import 'package:firebase_auth/firebase_auth.dart';
